Rename e2e test to pull chart tarball
