+++ integration/test/pullcharttarball/fixtures/test-chart/templates/pod.yaml
