+++ integration/test/pullcharttarball/fixtures/test-chart/templates/service.yaml
