+++ integration/test/pullcharttarball/fixtures/test-chart/values.yaml
