+++ integration/test/pullcharttarball/project.yaml
